+++ apps/www/src/app/introduction/page.tsx
@@ -1,5 +1,5 @@
 import ReactMarkdown from "react-markdown"
-import { DocsLinksGrid } from "~/client/DocsLink"
+import { DocsLinksGrid, getPageByHref } from "~/client/DocsLink"
 const markdown = `
 # **Welcome to Teampilot**
 
@@ -14,7 +14,9 @@ This means using Teampilot AI to programmatically generate text, data or media, 
 For example you could use Teampilot to generate a text for your web page or generate images for your blog posts.
 
 ## Integrate your product into Teampilot
-In Teampilot, there are things called **functions**. Functions are like small programs that Teampilot can use. For example, GPT isn't good at maths, so there's a function called **calculator** that GPT can use to do the math. Or a function that can visit a website.
+In Teampilot, there are things called [**functions**](${
+  getPageByHref("/functions")?.href
+}). Functions are like small programs that Teampilot can use. For example, GPT isn't good at maths, so there's a function called **calculator** that GPT can use to do the math. Or a function that can visit a website.
 
 With the Teampilot SDK, you can create and connect your own functions to Teampilot. So, for example, you could make a function that can get user data from your database and give it to Teampilot. Then, your Support team could ask Teampilot when a user named "John Doe" last paid. Teampilot would then call your function, your function would find the data from your database and give it back to Teampilot. Teampilot would then answer the question using the info from your function.
 `
